db-hackathon/requirements.txt
eth-brownie==1.21.0
fastapi==0.116.1